Refactoring osgi container
